refactor(broker): unify the three frontends into one infrabroker binary (#180)
The broker (HTTP+mTLS), mcp-broker (stdio MCP) and mcp-broker-http (OAuth HTTP
MCP) frontends were three thin mains over the same broker.Engine, differing only
in transport, yet every release artifact, Dockerfile stage, deploy unit and docs
page carried three names. Consolidate them into one `infrabroker` binary with the
transport as a subcommand — serve-http / serve-mcp / serve-mcp-http — with the
shared boot wiring (config load, engine, monitor listener, HTTPS serve) extracted
into internal/brokermain.

The legacy binaries remain as ~8-line DEPRECATED compat wrappers over the same
Run* functions, so existing MCP-client configs, systemd units, the container
ENTRYPOINT and the labs keep working unchanged — same flags, same behaviour, same
audit/metric labels. No security-semantics change: the engine, config and
per-transport defaults are untouched; server.json (which names no binary — it
appends -config to the image ENTRYPOINT) stays byte-identical.

Build/release: append infrabroker to the Makefile CMDS, one goreleaser build
entry + archive/docker ids, and one Dockerfile COPY. Deploy and labs need zero
edits (the wrappers preserve the cmd/<name> packages and -config-first CLI). The
cmd/broker and cmd/mcp-broker-http package tests moved to internal/brokermain;
docgen documents the routes under internal/brokermain (endpoints.md regenerated).
Docs updated wherever binary names are enumerated.

Closes #180
